WEBREF-53 :zap: Auto add jira info from branch name
